backend/services/tourist_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
qrcode==7.4.2
pillow==10.1.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1